Added consul configurations
