+++ FeedService/src/main/java/com/abc/feedservice/FeedServiceApplication.java
@@ -1,12 +1,24 @@
 package com.abc.feedservice;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
+@EnableDiscoveryClient
+@RestController
 public class FeedServiceApplication extends SpringBootServletInitializer {
+
+    ObjectMapper objectMapper = new ObjectMapper();
 
     @Override
     protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
@@ -15,6 +27,11 @@ public class FeedServiceApplication extends SpringBootServletInitializer {
 
     public static void main(String[] args) {
         SpringApplication.run(FeedServiceApplication.class, args);
+    }
+
+    @RequestMapping(path = "/", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public String getHealth() throws JsonProcessingException {
+        return objectMapper.writeValueAsString("{ \"status\" : \"healthy\"}");
     }
 }
