+++ FeedService/src/main/resources/application.properties
@@ -1,6 +1,12 @@
 # Feed Server Configurations
 spring.application.name=feed-service
 server.port=8080
+management.security.enabled=false
+
+#
+spring.cloud.consul.host=149.165.170.184
+spring.cloud.consul.port=8500
+spring.cloud.consul.discovery.health-check-path=/health
 
 # MongoDB Configurations
 spring.data.mongodb.host=localhost
