+++ FeedService/src/main/resources/bootstrap.yml
@@ -1,0 +1,12 @@
+spring:
+  application:
+    name: feed-service
+  cloud:
+    bus:
+      enabled: false
+    discovery:
+      enabled: false
+    consul:
+      enabled: false
+      config:
+        enabled: false
